+++ tests/loggers/test_neptune.py
@@ -22,7 +22,7 @@ from tests.helpers import BoringModel
 
 @patch('pytorch_lightning.loggers.neptune.neptune')
 def test_neptune_online(neptune):
-    logger = NeptuneLogger(api_key='test', project_name='project')
+    logger = NeptuneLogger(api_key='test', project='project')
 
     created_run = neptune.init()
 
@@ -38,7 +38,7 @@ def test_neptune_online(neptune):
 
 @patch('pytorch_lightning.loggers.neptune.neptune')
 def test_neptune_additional_methods(neptune):
-    logger = NeptuneLogger(api_key='test', project_name='project')
+    logger = NeptuneLogger(api_key='test', project='project')
 
     run_mock = MagicMock()
     neptune.init.return_value = run_mock
@@ -72,3 +72,54 @@ def test_neptune_leave_open_experiment_after_fit(neptune, tmpdir):
 
     logger_open_after_fit = _run_training(NeptuneLogger(offline_mode=True, close_after_fit=False))
     assert logger_open_after_fit.experiment.stop.call_count == 0
+
+
+def _assert_legacy(callable, *args, **kwargs):
+    try:
+        callable(*args, **kwargs)
+    except ValueError:
+        pass
+    else:
+        raise AssertionError(f"Should throw `ValueError`")
+
+
+def test_legacy_init_kwargs():
+    legacy_neptune_init_kwargs = [
+        'project_name',
+        'offline_mode',
+        'experiment_name',
+        'experiment_id',
+        'params',
+        'properties',
+        'upload_source_files',
+        'abort_callback',
+        'logger',
+        'upload_stdout',
+        'upload_stderr',
+        'send_hardware_metrics',
+        'run_monitoring_thread',
+        'handle_uncaught_exceptions',
+        'git_info',
+        'hostname',
+        'notebook_id',
+        'notebook_path',
+    ]
+    for legacy_kwarg in legacy_neptune_init_kwargs:
+        _assert_legacy(
+            lambda: NeptuneLogger(**{legacy_kwarg: None})
+        )
+
+
+def test_legacy_functions():
+    logger = NeptuneLogger(api_key='test', project='project')
+
+    # test all  functions
+    _assert_legacy(logger.log_metric)
+    _assert_legacy(logger.log_text)
+    _assert_legacy(logger.log_image)
+    _assert_legacy(logger.log_artifact)
+    _assert_legacy(logger.set_property)
+    _assert_legacy(logger.append_tags)
+
+    # test random args
+    _assert_legacy(logger.log_metric, 42, foo="bar")